refactor: move lozenges to their own file
